backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
networkx==3.2.1
flask-socketio==5.3.6
alembic==1.13.1
sqlalchemy==2.0.27
